# Fix JAX 0.7.0 compatibility: resolve static argument issues