--- a/mechafil_jax/supply.py
+++ b/mechafil_jax/supply.py
@@ -147,7 +147,7 @@ def update_cs_day(carry, x):
     )
     return (return_carry, None)
 
-@partial(jax.jit, static_argnums=(0,1,2,3,4,5,6))
+@partial(jax.jit, static_argnums=(0,1,2,6))
 def forecast_circulating_supply(
     start_date: np.datetime64,
     current_date: np.datetime64,
@@ -197,7 +197,7 @@ def forecast_circulating_supply(
     return cs_dict
 
 
-@partial(jax.jit, static_argnums=(0,1,2,3))
+@partial(jax.jit, static_argnums=(0,1))
 def initialise_circulating_supply_dict(
     start_date: np.datetime64,
     end_date: np.datetime64,