--- a/setup.py
+++ b/setup.py
@@ -10,12 +10,14 @@ setuptools.setup(
         "numpy", 
         "scipy", 
         "matplotlib",
-        "pystarboard @ git+https://git@github.com/protocol/pystarboard.git",
-    ],
+        "scenario-generator @ git+https://github.com/CELtd/scenario-generator", 
+        "pystarboard @ git+https://github.com/CELtd/pystarboard@updated_dependencies",
 
-    tests_require = [
-        'pytest',
-        'pandas==1.5.3',
-        'mechaFIL @ git+https://git@github.com/protocol/filecoin-mecha-twin.git@mechafil_jax'  # get the branch that is built for comparisons w/ jax
+    ],
+    extras_require = {
+    "test": [
+        "pytest",
+        "mechaFIL @ git+https://github.com/CELtd/filecoin-mecha-twin@mechafil-jax-pandas-updates",
     ]
+}
 )