# Fix JAX 0.7.0 compatibility: resolve static argument issues

## Summary
Fixes JAX 0.7.0 compatibility issues by resolving "Non-hashable static arguments" errors. JAX 0.7.0 introduced stricter validation for static arguments in JIT-compiled functions, requiring that all `static_argnums` be hashable at compile time. This PR removes dynamic JAX arrays from static argument lists while preserving truly static values like dates and integers.

## Error Resolved
```
ValueError: Non-hashable static arguments are not supported. An error occurred while trying to hash an object of type <class 'jax._src.interpreters.partial_eval.DynamicJaxprTracer'>, JitTracer<~int64[]>. The error was:
TypeError: unhashable type: 'DynamicJaxprTracer'
```

## Root Cause
Arguments marked as `static_argnums` must be compile-time constants, but several functions were incorrectly marking **dynamic JAX arrays from the data dictionary** as static arguments. JAX 0.7.0 requires static arguments to be hashable Python types, but values from `data[xxx]` are JAX tracers which are not hashable.

## Fix Strategy
Applied a simple, consistent rule across all functions:
- **Values from `data[xxx]`** → **Dynamic** (remove from `static_argnums`)
- **Non-data values** (dates, integers, strings) → **Static** (keep in `static_argnums`)

This eliminates the entire class of "data dictionary static argument" errors while preserving optimization opportunities for truly static values.

## Changes Made

All changes follow the same pattern: removing `data[xxx]` values from `static_argnums` while preserving truly static arguments.

### Summary of Changes

| File | Function | Static Args Removed | Kept Static |
|------|----------|-------------------|-------------|
| `vesting.py:23` | `compute_vesting_trajectory` | `start_vested_amt` (from `data["start_vested_amt"]`) | dates (0,1) |
| `minting.py:98` | `compute_baseline_power_array` | `init_baseline` (from `data["init_baseline_eib"]`) | dates (0,1) |
| `minting.py:34` | `compute_minting_trajectory_df` | `zero_cum_capped_power_eib` (from `data["zero_cum_capped_power_eib"]`) | dates (0,1), `minting_base` string (8) |
| `supply.py:150` | `forecast_circulating_supply` | `circ_supply_zero`, `locked_fil_zero`, `daily_burnt_fil` (from data dict) | dates (0,1,2), `duration` int (6) |
| `supply.py:200` | `initialise_circulating_supply_dict` | `circ_supply_zero`, `locked_fil_zero` (from data dict) | dates (0,1) |

### Detailed Changes

<details>
<summary>1. <code>mechafil_jax/vesting.py:23</code> - <code>compute_vesting_trajectory</code></summary>

```diff
-@partial(jax.jit, static_argnums=(0,1,2))
+@partial(jax.jit, static_argnums=(0,1))
 def compute_vesting_trajectory(
     start_date: np.datetime64, end_date: np.datetime64, start_vested_amt: int
 ) -> Dict:
```
</details>

<details>
<summary>2. <code>mechafil_jax/minting.py:98</code> - <code>compute_baseline_power_array</code></summary>

```diff
-@partial(jax.jit, static_argnums=(0,1,2))
+@partial(jax.jit, static_argnums=(0,1))
 def compute_baseline_power_array(
     start_date: np.datetime64, end_date: np.datetime64, init_baseline: float,
 ) -> Union[jnp.ndarray, NDArray, float]:
```
</details>

<details>
<summary>3. <code>mechafil_jax/minting.py:34</code> - <code>compute_minting_trajectory_df</code></summary>

```diff
-@partial(jax.jit, static_argnums=(0, 1, 6, 8))
+@partial(jax.jit, static_argnums=(0, 1, 8))
 def compute_minting_trajectory_df(
     # zero_cum_capped_power_eib removed from static (index 6)
     # minting_base string kept static (index 8)
```
</details>

<details>
<summary>4. <code>mechafil_jax/supply.py:150</code> - <code>forecast_circulating_supply</code></summary>

```diff
-@partial(jax.jit, static_argnums=(0,1,2,3,4,5,6))
+@partial(jax.jit, static_argnums=(0,1,2,6))
 def forecast_circulating_supply(
     # circ_supply_zero, locked_fil_zero, daily_burnt_fil removed (indices 3,4,5)
     # dates and duration kept static (0,1,2,6)
```
</details>

<details>
<summary>5. <code>mechafil_jax/supply.py:200</code> - <code>initialise_circulating_supply_dict</code></summary>

```diff
-@partial(jax.jit, static_argnums=(0,1,2,3))
+@partial(jax.jit, static_argnums=(0,1))
 def initialise_circulating_supply_dict(
     # circ_supply_zero, locked_fil_zero removed (indices 2,3)
     # dates kept static (0,1)
```
</details>

## Static Argument Rules Applied

### ✅ **Should be Static:**
- `np.datetime64` objects (for array indexing and length calculations)
- Integer constants like `duration`, `forecast_length`
- String parameters like `minting_base`

### ❌ **Should NOT be Static:**
- Any value from `data` dictionary (e.g., `data["start_vested_amt"]`)
- JAX arrays (`jnp.array`)
- Values that need to be differentiable
- Runtime-computed values

## Performance Impact

Comprehensive performance testing comparing JAX 0.4.26 vs 0.7.0 on a 20-year simulation (7,300 days):

| Metric | JAX 0.4.26 (Main) | JAX 0.7.0 (This PR) | Difference |
|--------|-------------------|---------------------|------------|
| **Average Time** | 14ms | 36ms | +22ms |
| **Min Time** | 13ms | 34ms | +21ms |
| **Max Time** | 15ms | 38ms | +23ms |
| **Std Deviation** | <1ms | 2ms | Slightly more variance |

**Results**: The absolute difference is 22ms for a 20-year simulation, and compatibility benefits outweigh the minor performance cost.

## Alternative Performance-Optimized Solution

If the ~22ms performance difference is critical, there's an alternative approach that would restore original performance:

### **Convert Data Dictionary Values to Python Scalars**

Instead of making `data[xxx]` values dynamic, convert them to Python scalars in the calling code:

```python
# In sim.py, instead of:
start_vested_amt = data["start_vested_amt"]  # JAX array (dynamic)

# Do this:
start_vested_amt = float(data["start_vested_amt"])  # Python scalar (can be static)
```

Then revert `static_argnums` back to original values:
```python
@partial(jax.jit, static_argnums=(0,1,2))  # Back to original
def compute_vesting_trajectory(start_date, end_date, start_vested_amt):
```
### **Critical Consideration: Parameter Variation**

**If these values change between simulations** (different scenarios, parameter sweeps, optimization), the scalar approach becomes problematic:

```python
# ❌ This won't work well with static scalars:
for scenario in scenarios:
    data = load_scenario_data(scenario)  # Different values each time
    start_vested_amt = float(data["start_vested_amt"])  # Forces recompilation!
    result = compute_vesting_trajectory(..., start_vested_amt, ...)
    # JAX recompiles function for each different scalar value
```

**If values vary across simulations, static scalars could actually be SLOWER** due to constant recompilation.

## Testing

- ✅ All simulations run without errors on JAX 0.7.0
- ✅ Numerical results verified to be identical

## Files Changed

- `mechafil_jax/vesting.py` - 1 function
- `mechafil_jax/minting.py` - 2 functions  
- `mechafil_jax/supply.py` - 2 functions